Refactored curve names
