+++ group/p256/doc.go
@@ -1,3 +1,3 @@
 // Package nist implements cryptographic groups and ciphersuites
 // based on the NIST standards, using Go's built-in crypto library.
-package nist
+package p256
