+++ group/p256/group_test.go
@@ -1,4 +1,4 @@
-package nist
+package p256
 
 import (
 	"testing"
