+++ group/p256/p256.go
@@ -1,4 +1,4 @@
-package nist
+package p256
 
 import (
 	"crypto/elliptic"
