+++ group/p256/qrsuite.go
@@ -1,4 +1,4 @@
-package nist
+package p256
 
 import (
 	"crypto/cipher"
