+++ group/p256/suite.go
@@ -1,4 +1,4 @@
-package nist
+package p256
 
 import (
 	"crypto/cipher"
