+++ group/var_ed25519/basic.go
@@ -1,7 +1,7 @@
 //go:build experimental
 // +build experimental
 
-package curve25519
+package var_ed25519
 
 import (
 	"crypto/cipher"
