+++ group/var_ed25519/basic_test.go
@@ -1,7 +1,7 @@
 //go:build experimental
 // +build experimental
 
-package curve25519
+package var_ed25519
 
 import (
 	"testing"
