+++ group/var_ed25519/curve_test.go
@@ -1,4 +1,4 @@
-package curve25519
+package var_ed25519
 
 import (
 	"testing"
