+++ proof/proof_test.go
@@ -7,9 +7,9 @@ import (
 	"testing"
 
 	"go.dedis.ch/kyber/v3"
-	"go.dedis.ch/kyber/v3/group/curve25519"
 	"go.dedis.ch/kyber/v3/group/edwards25519"
-	"go.dedis.ch/kyber/v3/group/nist"
+	"go.dedis.ch/kyber/v3/group/p256"
+	"go.dedis.ch/kyber/v3/group/var_ed25519"
 	"go.dedis.ch/kyber/v3/xof/blake2xb"
 )
 
@@ -256,10 +256,10 @@ func BenchmarkProof(b *testing.B) {
 		Suite
 	}{
 		{edwards25519.NewBlakeSHA256Ed25519()},
-		{curve25519.NewBlakeSHA256Curve25519(false)},
-		{curve25519.NewBlakeSHA256Curve25519(true)},
-		{nist.NewBlakeSHA256P256()},
-		{nist.NewBlakeSHA256QR512()},
+		{var_ed25519.NewBlakeSHA256Curve25519(false)},
+		{var_ed25519.NewBlakeSHA256Curve25519(true)},
+		{p256.NewBlakeSHA256P256()},
+		{p256.NewBlakeSHA256QR512()},
 	}
 
 	for _, suite := range suites {
