+++ shuffle/vartime_test.go
@@ -3,29 +3,29 @@ package shuffle
 import (
 	"testing"
 
-	"go.dedis.ch/kyber/v3/group/nist"
+	"go.dedis.ch/kyber/v3/group/p256"
 )
 
 func BenchmarkBiffleP256(b *testing.B) {
-	biffleTest(nist.NewBlakeSHA256P256(), b.N)
+	biffleTest(p256.NewBlakeSHA256P256(), b.N)
 }
 
 func Benchmark2PairShuffleP256(b *testing.B) {
-	pairShuffleTest(nist.NewBlakeSHA256P256(), 2, b.N)
+	pairShuffleTest(p256.NewBlakeSHA256P256(), 2, b.N)
 }
 
 func Benchmark10PairShuffleP256(b *testing.B) {
-	pairShuffleTest(nist.NewBlakeSHA256P256(), 10, b.N)
+	pairShuffleTest(p256.NewBlakeSHA256P256(), 10, b.N)
 }
 
 func Benchmark2Pair2SeqShuffleP256(b *testing.B) {
-	sequenceShuffleTest(nist.NewBlakeSHA256P256(), 2, 2, b.N)
+	sequenceShuffleTest(p256.NewBlakeSHA256P256(), 2, 2, b.N)
 }
 
 func Benchmark2Pair10SeqShuffleP256(b *testing.B) {
-	sequenceShuffleTest(nist.NewBlakeSHA256P256(), 2, 10, b.N)
+	sequenceShuffleTest(p256.NewBlakeSHA256P256(), 2, 10, b.N)
 }
 
 func Benchmark10Pair10SeqShuffleP256(b *testing.B) {
-	sequenceShuffleTest(nist.NewBlakeSHA256P256(), 10, 10, b.N)
+	sequenceShuffleTest(p256.NewBlakeSHA256P256(), 10, 10, b.N)
 }
