+++ suites/all.go
@@ -2,7 +2,7 @@ package suites
 
 import (
 	"go.dedis.ch/kyber/v3/group/edwards25519"
-	"go.dedis.ch/kyber/v3/group/nist"
+	"go.dedis.ch/kyber/v3/group/p256"
 	"go.dedis.ch/kyber/v3/pairing/bn256"
 	"go.dedis.ch/kyber/v3/pairing/circl_bls12381"
 )
@@ -10,8 +10,8 @@ import (
 func init() {
 	// Those are variable time suites that shouldn't be used
 	// in production environment when possible
-	register(nist.NewBlakeSHA256P256())
-	register(nist.NewBlakeSHA256QR512())
+	register(p256.NewBlakeSHA256P256())
+	register(p256.NewBlakeSHA256QR512())
 	register(bn256.NewSuiteG1())
 	register(bn256.NewSuiteG2())
 	register(bn256.NewSuiteGT())
